Split job commands into their own source file.
